fes/requirements.txt
Django==2.0.8
django-bower==5.2.0
mysqlclient==1.3.13
pytz==2018.5
schedule==0.5.0
six==1.11.0
django-phonenumber-field==2.0.1
django-datetime-widget==0.9.3
django-filter==2.0.0